flask_app/requirements.txt
Flask
sagemaker
boto3
python-dotenv